# Conflicts:
#	Angular_Final_Project/Ang_Final_Project/src/app/app.component.ts
#	Angular_Final_Project/Ang_Final_Project/src/app/app.config.ts
#	Angular_Final_Project/Ang_Final_Project/src/app/app.routes.ts
#	Angular_Final_Project/Ang_Final_Project/src/app/components/home/home.component.html
#	Angular_Final_Project/Ang_Final_Project/src/app/components/home/home.component.ts

+++ src/main/java/generation/italy/org/ravenclaw/models/services/JpaUtenteService.java
@@ -1,14 +1,17 @@
 
 package generation.italy.org.ravenclaw.models.services;
 
-import generation.italy.org.ravenclaw.exceptions.EntityNotFoundException;
-import generation.italy.org.ravenclaw.models.entities.Autore;
-import generation.italy.org.ravenclaw.models.entities.Libro;
+import generation.italy.org.ravenclaw.models.dtos.request.PasswordUpdateRequest;
+import generation.italy.org.ravenclaw.models.dtos.response.UtenteResponse;
+import generation.italy.org.ravenclaw.models.entities.Authority;
 import generation.italy.org.ravenclaw.models.entities.Utente;
-import generation.italy.org.ravenclaw.models.repositories.AutoreRepository;
 import generation.italy.org.ravenclaw.models.repositories.UtenteRepository;
-import org.springframework.http.ResponseEntity;
+import generation.italy.org.ravenclaw.models.services.utils.FindAuthenticatedUser;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,10 +19,28 @@ import java.util.Optional;
 @Service
 public class JpaUtenteService implements UtenteService{
     private UtenteRepository utenteRepo;
+    private final FindAuthenticatedUser findAuthenticatedUser;
+    private final PasswordEncoder passwordEncoder;
 
-    public JpaUtenteService(UtenteRepository utenteRepo) {
+
+    public JpaUtenteService(UtenteRepository utenteRepo, FindAuthenticatedUser findAuthenticatedUser, PasswordEncoder passwordEncoder) {
         this.utenteRepo = utenteRepo;
+        this.findAuthenticatedUser = findAuthenticatedUser;
+        this.passwordEncoder = passwordEncoder;
     }
+
+    @Transactional(readOnly = true)
+    @Override
+    public UtenteResponse getUserInfo() {
+        Utente utente = findAuthenticatedUser.getAuthenticatedUser();
+        return new UtenteResponse(
+                utente.getUtenteId(),
+                utente.getNome(),
+                utente.getEmail(),
+                utente.getAuthorities().stream().map(auth -> (Authority) auth).toList()
+        );
+    }
+
 
     @Override
     public Optional<Utente> findUtenteById(int id)  {
@@ -45,9 +66,62 @@ public class JpaUtenteService implements UtenteService{
 
     @Override
     public boolean deleteUtente(int id) {
+        Utente utente = findAuthenticatedUser.getAuthenticatedUser();
+        if(isLastAdmin(utente)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Un Admin non può cancellare se stesso");
+        }
         utenteRepo.deleteById(id);
         return true;
     }
+
+    @Transactional
+    @Override
+    public void updatePassword(PasswordUpdateRequest passwordUpdateRequest) {
+        Utente user = findAuthenticatedUser.getAuthenticatedUser();
+
+        if (!isOldPasswordCorrect(user.getPassword(), passwordUpdateRequest.getVecchiaPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password attuale errata");
+        }
+
+        if (!isNewPasswordConfirmed(passwordUpdateRequest.getNewPassword(),
+                passwordUpdateRequest.getNewPasswordRepeat())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Le nuove password non coincidono");
+        }
+
+        if (!isNewPasswordDifferent(passwordUpdateRequest.getVecchiaPassword(),
+                passwordUpdateRequest.getNewPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "La vecchia password e la nuova password non possono essere uguali");
+        }
+
+        user.setPassword(passwordEncoder.encode(passwordUpdateRequest.getNewPassword()));
+        utenteRepo.save(user);
+    }
+
+    private boolean isOldPasswordCorrect(String currentPassword, String oldPassword) {
+        return passwordEncoder.matches(oldPassword, currentPassword);
+    }
+
+    private boolean isNewPasswordConfirmed(String newPassword, String newPasswordConfirmation) {
+        return newPassword.equals(newPasswordConfirmation);
+    }
+
+    private boolean isNewPasswordDifferent(String oldPassword, String newPassword) {
+        return !oldPassword.equals(newPassword);
+    }
+
+    private boolean isLastAdmin(Utente user) {
+        boolean isAdmin = user.getAuthorities().stream()
+                .anyMatch(authority -> "ROLE_ADMIN".equals(authority.getAuthority()));
+
+        if (isAdmin) {
+            long adminCount = utenteRepo.countAdminUsers();
+            return adminCount <= 1;
+        }
+
+        return false;
+    }
+
+
 
 }
